Split out pip property functionality from PipRetryInfo into PipPropertyInfo
